move .results to race
